+++ scripts/generator.js
@@ -4,8 +4,11 @@
 
 // using https://unpkg.com/@ionic/docs@6.6.0/core.json
 // using https://unpkg.com/@ionic/docs@7.0.2/core.json
+// using https://unpkg.com/@ionic/docs@7.0.3/core.json
 
 const fs = require("fs");
+const https = require('https');
+
 
 const kebabize = str => {
   return str.split('').map((letter, idx) => {
@@ -23,10 +26,12 @@ function clearAndUpper(text) {
   return text.replace(/-/, "").toUpperCase();
 }
 
-// load static
-const coreJson = require("./core.json");
+
 
 const doStuff = () => {
+  // load static
+  const coreJson = require("./core.json");
+
   var dir = "./generated";
 
   if (!fs.existsSync(dir)) {
@@ -161,7 +166,7 @@ export function testUserAgent(win: Window, expr: RegExp);
       componentTypes = componentTypes + `interface ${tagAsPascal} {\n`; //  extends EventTarget
 
       // let's dump the props
-      console.log('has props', props);
+      // console.log('has props', props);
       props.forEach(prop => {
 
         //   "disabled"?: boolean;
@@ -175,7 +180,7 @@ export function testUserAgent(win: Window, expr: RegExp);
       })
 
       // let's dump the events
-      console.log('has props', events);
+      // console.log('has props', events);
       events.forEach(event => {
 
         //     "on:ionSlideReachEnd"?: () => void;
@@ -248,6 +253,135 @@ export function testUserAgent(win: Window, expr: RegExp);
   );
 };
 
-doStuff();
+function download(res) {
+  let data = '';
+  res.on('data', (chunk) => {
+    data += chunk;
+  });
+  res.on('end', () => {
+    fs.writeFile(file, data, (err) => {
+      if (err) throw err;
+      console.log(`File saved as ${file} - length ${data.length}`);
+      doStuff();
+    });
+  });
+}
+
+function extractVersion(url) {
+  const regex = /@(\d+\.\d+\.\d+)/;
+  const match = url.match(regex);
+
+  if (match) {
+    const version = match[1];
+    // console.log(version); // Output: "7.0.3"
+    return version;
+  } else {
+    console.log('No version number found');
+    return 'ERROR'
+  }
+}
+
+function bumpPackageJson(version) {
+
+  const currentDirectory = process.cwd();
+
+  if (currentDirectory.includes("script")) {
+    console.log("Current directory contains the word 'script'");
+  } else {
+    console.log("Current directory does not contain the word 'script'");
+    process.exit()
+  }
+
+  const filePath = '../package.json';
+  const packageName = 'ionic-svelte';
+
+  fs.readFile(filePath, 'utf-8', (err, data) => {
+    if (err) throw err;
+
+    // make backup
+    fs.writeFile(filePath + '.bak', data, 'utf-8', (err) => {
+      if (err) throw err;
+      console.log(`Backup created of ${filePath}`);
+    });
+
+    const packageJson = JSON.parse(data);
+
+    if (packageJson.dependencies && packageJson.dependencies[packageName]) {
+      packageJson.dependencies[packageName] = version;
+    } else if (packageJson.devDependencies && packageJson.devDependencies[packageName]) {
+      packageJson.devDependencies[packageName] = version;
+    } else {
+      console.log(`Package ${packageName} not found in dependencies or devDependencies`);
+    }
+
+    const updatedPackageJson = JSON.stringify(packageJson, null, 2);
+
+    fs.writeFile(filePath, updatedPackageJson, 'utf-8', (err) => {
+      if (err) throw err;
+      console.log(`Version of package ${packageName} updated to ${version}\n${updatedPackageJson}`);
+    });
+  });
+
+}
 
 
+const url = 'https://unpkg.com/@ionic/docs/core.json';
+const file = 'core.json';
+
+https.get(url, (res) => {
+  if (res.statusCode >= 300 && res.statusCode <= 399 && res.headers.location) {
+    // Redirect detected
+
+    console.log('redirect', res.headers.location)
+    const version = extractVersion(res.headers.location);
+    console.log('Version found', version)
+
+    bumpPackageJson(version);
+    // bumpCreatorPackages(version, '../packages/create-capacitor-svelte-app/src/creator.js');
+    // bumpCreatorPackages(version,'../packages/create-ionic-svelte-app/src/creator.js');
+
+    https.get('https://unpkg.com' + res.headers.location, (res) => {
+      download(res);
+    });
+  } else {
+    download(res);
+  }
+});
+
+function bumpCreatorPackages(version, fileName) {
+  // Read the contents of the creator.js file
+
+
+
+  fs.readFile(fileName, 'utf8', (err, data) => {
+    if (err) {
+      console.error(err);
+      return;
+    }
+
+    // make backup
+    fs.writeFile(fileName + '.bak', data, 'utf-8', (err) => {
+      if (err) throw err;
+      console.log(`Backup created of ${fileName}`);
+    });
+
+    // Define the search and replace strings
+    const searchStr = "['@ionic/core@";
+    const replaceStr = "['@ionic/core@" + version;
+
+    // Use a regular expression to search and replace the desired string
+    const regex = new RegExp(`${searchStr}\\d+\\.\\d+\\.\\d+'`, 'g');
+    const newData = data.replace(regex, replaceStr);
+
+    // Write the updated contents back to the file
+    fs.writeFile('creasssstor.js', newData, 'utf8', (err) => {
+      if (err) {
+        console.error(err);
+        return;
+      }
+
+      console.log('File updated successfully!');
+    });
+  });
+
+}