+++ CHANGELOG.md
@@ -1,6 +1,11 @@
 # Change Log Ionic-Svelte-NPM
 All notable changes to this project will be documented in this file.
 
+## 0.5.77
+- Bumped to 7.0.3 - including creator scripts
+- Fix on MenuI
+- IonPage.svelte now holds a secret
+
 ## 0.5.74/75/76 Ionic 7 (Ionic 7.0.2)
 - Ionic 7 package landed - demo app is not upgraded yet
 - Creator scripts are updated to 7.0.2 as well - fixing to that version as I assume there will be updates on core